refactor(ws): split ws_handlers god file into a package
Decompose the 1932-line custom_components/beatify/server/ws_handlers.py
into a cohesive ws_handlers/ package, re-exporting every public name from
__init__.py so all existing `from ...ws_handlers import X` imports keep
working. Strictly behavior-preserving — handler bodies, dispatch wiring and
public names are unchanged; only their physical location moves.

Module map:
- _helpers.py  : state-redaction + admin-auth helpers
- lifecycle.py : join/reconnect/leave/state/ping/onboarding/reaction/timeout
- admin.py     : admin connect + admin action dispatch table + sub-handlers
- guessing.py  : submit/steal/artist/movie/title-artist guess+vote+override
- __init__.py  : re-exports + data-quality report handler (#911/#1384), kept
                 here because its tests patch ws_handlers.async_get_clientsession
                 /aiohttp/asyncio on this package namespace.

Fixes #1588
